Add Polyphony
Add simple polyphony to the following classes

- The simpleOSC (4 modules)
- The plusOSC (4 modules)
- The simplex oscillator (1 module)
- The logic and math modules

Implement polyphony the fairly naive way; trigger on FREQ_CV (or in
the math cases INPUT_A) for size and use getPolyVoltage everywhere
else. Do not do any inter-voice SIMD or the like; just make multiple
isntances of the calculators.

Unresolved is how to address the scope in the simplex operator; for
now it shows just voice 0.