Fix some scope / Framebuffer problems (#17)
The wave oscillator polyphony now works. Some occasional crashes
emerged from the scope for a couple of reasons. First I added
a few defensive null checks, but far more importantly, I used
the FramebufferWidget more in the way it was intended, namely
including an OpaqueWidget which dirties the parent appropriately.
Some of the log messages about the FBW having size (nan,nan) and
definitely the occasional crashes in exit in a widget destructor
were caused by the prior pattern, I think.

Anyway with this change it all seems super stable to me.
Best!